Tests moved to class _BuildHistoryTest, The code and methods were refactored, Deleted old class.
